pyproject.toml
[tool.poetry]
name = "pycontainer"
version = "0.1.0"
description = ""
authors = ["Keith Nordstrom <keith@timeli.io>"]
readme = "README.rst"
packages = [{include = "library"}, {include = "processor"}]

[tool.poetry.scripts]
processor = "processor.main:start"
event_processor = "processor.main:start_kafka_calendar"
document_processor = "processor.main:start_kafka_documents"
employee_processor = "employees.main:start"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^8.1.1"
google-api-python-client = "^2.122.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"
weaviate-client = "^4.6.2"
langchain = "^0.1.12"
langchain-experimental = "^0.0.55"
langchain-openai = "^0.0.8"
beautifulsoup4 = "^4.8.2"
email-reply-parser = "^0.5.12"
coverage = "^7.4.4"
kafka-python = "^2.0.2"
python-dotenv = "^1.0.1"
ics = "^0.7.2"
neo4j = "^5.20.0"
langchain-community = "^0.0.36"
matplotlib = "^3.8.4"
numpy = "^1.26.4"
groq = "^0.5.0"
langchain-groq = "^0.1.3"
python-docx = "^1.1.2"
pypdf = "^4.2.0"
textextract = "^1.2.0"
textract = "^1.6.5"
slack-sdk = "^3.27.2"
pytest-docker = "^3.1.1"
icalendar = "^5.0.12"
recurring-ical-events = "^2.2.1"
fastapi = "^0.111.0"
uvicorn = "^0.30.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
